rooter ayarlamları yapıldı içerik düzenlecek
