+++ ilan-uygulama/src/features/filterSlice.js
@@ -1,5 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { is_ilanlari_veriler } from "../components/IlanList";
+import { is_ilanlari_veriler } from "../components/IlanKartları";
 
 export const filterSlice = createSlice({
     name: 'filter',
@@ -10,7 +10,8 @@ export const filterSlice = createSlice({
         filter: (state, action) => {
             state.items = state.items.filter(item =>
                 item.location.toLowerCase().includes(action.payload.toLowerCase()) ||
-                item.job_title.toLowerCase().includes(action.payload.toLowerCase())
+                item.job_title.toLowerCase().includes(action.payload.toLowerCase()) ||
+                item.category.toLowerCase().includes(action.payload.toLowerCase())
             );
         },
     },
